Fix: browsers don't get generic messages.
Sinapse sets the channel as the event name, which has the problem on
browsers to generate specific events, with no possible way to catch
all events (`onmessage` is only called when the `event:` line is
skipped).

Since this forces frontend code to specifically listen to specified
events, this has been disabled for now. It should come back as a
global setting and maybe as a request parameter.